Trialwebsite/app/requirements.txt
pytest
requests
flask
flask-mysqldb
flask_socketio
passlib
mysqlclient
flask.session
